requirements.txt
psycopg2-binary
pyyaml
pysimplegui